refactor: consolidate window.env calls to single file
