+++ app/components/SocketProvider.js
@@ -13,6 +13,7 @@ import PoliciesStore from "stores/PoliciesStore";
 import ViewsStore from "stores/ViewsStore";
 import AuthStore from "stores/AuthStore";
 import UiStore from "stores/UiStore";
+import env from "env";
 
 export const SocketContext: any = React.createContext();
 
@@ -34,7 +35,7 @@ class SocketProvider extends React.Component<Props> {
   @observable socket;
 
   componentDidMount() {
-    if (!window.env.WEBSOCKETS_ENABLED) return;
+    if (!env.WEBSOCKETS_ENABLED) return;
 
     this.socket = io(window.location.origin, {
       path: "/realtime",
