+++ app/env.js
@@ -1,0 +1,3 @@
+// @flow
+const env = window.env;
+export default env;
