+++ app/index.js
@@ -10,6 +10,7 @@ import ScrollToTop from "components/ScrollToTop";
 import Toasts from "components/Toasts";
 import Theme from "components/Theme";
 import Routes from "./routes";
+import env from "env";
 
 let DevTools;
 if (process.env.NODE_ENV !== "production") {
@@ -44,7 +45,7 @@ if (element) {
 window.addEventListener("load", async () => {
   // installation does not use Google Analytics, or tracking is blocked on client
   // no point loading the rest of the analytics bundles
-  if (!window.env.GOOGLE_ANALYTICS_ID || !window.ga) return;
+  if (!env.GOOGLE_ANALYTICS_ID || !window.ga) return;
 
   // https://github.com/googleanalytics/autotrack/issues/137#issuecomment-305890099
   await import("autotrack/autotrack.js");
