+++ app/utils/domains.js
@@ -1,0 +1,7 @@
+// @flow
+import { stripSubdomain } from "shared/utils/domains";
+import env from "env";
+
+export function getCookieDomain(domain: string) {
+  return env.SUBDOMAINS_ENABLED ? stripSubdomain(domain) : domain;
+}
