+++ server/middlewares/authentication.js
@@ -3,10 +3,10 @@ import JWT from "jsonwebtoken";
 import { type Context } from "koa";
 import { User, ApiKey } from "../models";
 import { getUserForJWT } from "../utils/jwt";
+import { getCookieDomain } from "../utils/domains";
 import { AuthenticationError, UserSuspendedError } from "../errors";
 import addMonths from "date-fns/add_months";
 import addMinutes from "date-fns/add_minutes";
-import { getCookieDomain } from "../../shared/utils/domains";
 
 export default function auth(options?: { required?: boolean } = {}) {
   return async function authMiddleware(ctx: Context, next: () => Promise<*>) {
